requirements.txt
fastapi
uvicorn
pytesseract
Pillow
numpy
opencv-python
matplotlib
nltk
pdf2image
python-multipart